Temp, Humid and water sensor added
Sensors added. observing all so sorted, will now check the front  not just left and right.

Next steps is to sort out the IR senor on ESP as is either broken or refusing to work.  Also to work on feedback given on section 4(Bulletpoint tests) which are to be conducted tomorrow hopefully.