multicore wifi and bluetooth created
the use od dual core works well but are a lot of issues that need fixing/testing.
Documentation has been starter.
Connection lost need to be improved! either needs to be set as an interrupt, if this is to complex for the interrupt set up then just reset the board. Would be nice to have a function where nothing happens until the board and the device is paired and not just the board being ready, set all pins OFF. if need be.