wifi ip and wifi server reader
wifi ip- uses esp32 as a station and access point where a client has been created and a 'menu' is used to control the buggy. This has to many issues to deal with however- there is no feeback from the buggy. It is not communication to the buggy clearly.

Wifi server reader works well- it communicates to the buggy clearly using the IP address and port number. Feedback works but not well. this needs fixing.